requirements.txt
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
idna==2.7
nose==1.3.7
python-dotenv==0.8.2
requests==2.19.1
requests-mock==1.5.2
six==1.11.0
urllib3==1.23
